#39 Move supervision constants to top level, make clientTag.Id long

+++ c2mon-server/c2mon-server-client/src/test/java/cern/c2mon/server/client/publish/TagValuePublisherTest.java
@@ -190,7 +190,7 @@ public class TagValuePublisherTest {
     synchronized (updateLock) {
       assertNotNull(update); //message or update were null
       assertEquals(tag.getValue(), update.getValue());
-      assertEquals(tag.getId(), update.getId());
+      assertEquals(tag.getId(), update.getId().longValue());
 //      assertEquals(alarms.size(), update.getAlarms().size());
       assertEquals(tag.getDescription(), update.getDescription());
       assertEquals(tag.getValueDescription(), update.getValueDescription());
